feat: add confirm dialog befroe delete project

+++ src/routes/(app)/project-management/+page.svelte
@@ -9,13 +9,16 @@
 	import type { ErrorResponse } from '../../../lib/interface/error-response';
 	import Button, { Label } from '@smui/button';
 	import { onMount } from 'svelte';
+	import Dialog, { Title, Content, Actions } from '@smui/dialog';
 	import toast from 'svelte-french-toast';
 	import { userStore } from '../../../lib/store/user.js';
 
 	let loading = false;
 	let loadingDialog = true;
 	let openCreateDialog = false;
+	let openDeleteDialog = false;
 	let projectList: Project[] = [];
+	let selectedProjectForDelete: Project | undefined;
 
 	let projectName = '';
 
@@ -40,25 +43,35 @@
 		});
 	}
 
-	async function deleteProject(project: Project) {
-		loading = false;
-		const token: string = $userStore.token || '';
-		await fetch(`api/project/delete/${project.projectId}`, {
-			method: 'DELETE',
-			headers: new Headers({ Authorization: `Bearer ${token}` })
-		}).then(async (response: Response) => {
-			if (!response.ok) {
-				const error: ErrorResponse = await response.json();
-				toast.error(error.detail);
-			} else {
-				toast.success(`Delete project ${project.projectName} successfully."`);
-				await loadProject();
-			}
-		});
+	async function deleteProject(project: Project | undefined) {
+		if (project) {
+			loading = false;
+			const token: string = $userStore.token || '';
+			await fetch(`api/project/delete/${project.projectId}`, {
+				method: 'DELETE',
+				headers: new Headers({ Authorization: `Bearer ${token}` })
+			}).then(async (response: Response) => {
+				if (!response.ok) {
+					const error: ErrorResponse = await response.json();
+					toast.error(error.detail);
+				} else {
+					toast.success(`Delete project ${project.projectName} successfully.`);
+					await loadProject();
+				}
+				selectedProjectForDelete = undefined;
+			});
+		} else {
+			toast.error('No project selected.');
+		}
 	}
 
-	function toggleDialog(): void {
+	function toggleCreateDialog(): void {
 		openCreateDialog = !openCreateDialog;
+	}
+
+	function toggleDeleteDialog(project: Project) {
+		selectedProjectForDelete = project;
+		openDeleteDialog = true;
 	}
 </script>
 
@@ -66,7 +79,12 @@
 	<Card style="padding: 20px;">
 		<div style="display: flex; justify-content: space-between">
 			<h2 class="mdc-typography--headline6" style="margin: 0;">Project Management</h2>
-			<Button disabled={!loading} color="secondary" variant="unelevated" on:click={toggleDialog}>
+			<Button
+				disabled={!loading}
+				color="secondary"
+				variant="unelevated"
+				on:click={toggleCreateDialog}
+			>
 				<Icon path={mdiFolderPlus} />
 				<Label style="margin-left: 5px">Create project</Label>
 			</Button>
@@ -91,7 +109,12 @@
 									<Icon path={mdiFolderEdit} />
 									<Label style="margin-left: 5px">Edit</Label>
 								</Button>
-								<Button disabled={!loading} on:click={deleteProject(project)}>
+								<Button
+									disabled={!loading}
+									on:click={() => {
+										toggleDeleteDialog(project);
+									}}
+								>
 									<Icon path={mdiDeleteEmpty} />
 									<Label style="margin-left: 5px">Delete</Label>
 								</Button>
@@ -111,6 +134,30 @@
 	</Card>
 
 	<ProjectDialog bind:loadingDialog bind:openCreateDialog bind:projectName />
+	<Dialog
+		bind:open={openDeleteDialog}
+		scrimClickAction=""
+		escapeKeyAction=""
+		aria-labelledby="mandatory-title"
+		aria-describedby="mandatory-content"
+	>
+		<Title id="mandatory-title">Project Delete Confirmation</Title>
+		<Content id="mandatory-content">
+			Are you sure to delete project {selectedProjectForDelete?.projectName || 'N/A'}?
+		</Content>
+		<Actions>
+			<Button action="close">
+				<Label>Close</Label>
+			</Button>
+			<Button
+				color="secondary"
+				variant="unelevated"
+				on:click={deleteProject(selectedProjectForDelete)}
+			>
+				<Label>Confirm</Label>
+			</Button>
+		</Actions>
+	</Dialog>
 </div>
 
 <style lang="css">
